requirements.txt
aiohttp>=3.9.0
anyio==4.12.0
certifi==2025.11.12
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
python-dotenv==1.2.1
python-telegram-bot[job-queue]==21.11.1
pywin32==311

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0
